+++ changelog.md
@@ -8,6 +8,8 @@
 
 - ``re.split`` for empty regular expressions now yields every character in
   the string which is what other programming languages chose to do.
+- The returned tuple of ``system.instantiationInfo`` now has a third field
+  containing the column of the instantiation.
 
 - ``cookies.setCookie` no longer assumes UTC for the expiration date.
 
@@ -54,3 +56,4 @@
 
 - The VM's instruction count limit was raised to 1 billion instructions in order
   to support more complex computations at compile-time.
+- Added ``macros.getProjectPath`` and ``ospaths.putEnv`` procs to VM.